refactor: [BE] 레드박스 컨트롤러 분리
